fix: rename components to pascal case
